#595 Standalone HJD conversion plugin #456 Plug-ins: HJD #376 HJD conversion for file-based formats
Initial cut of HJD conversion plugin. This converts any obs that are not in HJD based upon observation retriever via which observation was loaded. A question from Mike Miller prompted this plugin. Two refinements are needed: 1. show list of loaded datasets that are not yet HJD, allowing selection via checkboxes; 2. populate RA and Dec with coords from one of these datasets (the first) if available; the user can override this.